backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
gunicorn==23.0.0
youtube-transcript-api==0.6.3
google-api-python-client==2.159.0
supabase==2.11.0
python-dotenv==1.0.1
requests==2.32.3
